Move query into pipeline since the separation was pointless
